+++ CHANGELOG.md
@@ -23,6 +23,9 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian/aws-infrastructure/compare/release-2.16.0...master
 
+### Deprecated
+- Replace `network` method with `instanceType` in `Ec2VirtualUsersFormula.Builder`.
+
 ## [2.16.0] - 2019-09-03
 [2.16.0]: https://github.com/atlassian/aws-infrastructure/compare/release-2.15.0...release-2.16.0