+++ CHANGELOG.md
@@ -21,7 +21,10 @@ Adding a requirement of a major version of a dependency is breaking a contract.
 Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
-[Unreleased]: https://github.com/atlassian/aws-infrastructure/compare/release-2.17.0...master
+[Unreleased]: https://github.com/atlassian/aws-infrastructure/compare/release-2.18.0...master
+
+## [2.18.0] - 2019-10-17
+[2.18.0]: https://github.com/atlassian/aws-infrastructure/compare/release-2.17.0...release-2.18.0
 
 ### Added
 - Add `UriJiraFormula`. Resolve [JPERF-579].